requirements.txt
altair
numpy
pandas
pydeck
streamlit
google-generativeai
python-dotenv